Extract common binding utils
