+++ src/js/CMakeLists.txt
@@ -50,6 +50,7 @@ target_include_directories(${PROJECT_NAME} PRIVATE
   "${node-api-headers_SOURCE_DIR}/include"
   "${node-addon-api_SOURCE_DIR}"
   "${CMAKE_CURRENT_SOURCE_DIR}"
+  "${CMAKE_CURRENT_SOURCE_DIR}/.."
 )
 
 target_link_libraries(${PROJECT_NAME} PRIVATE openvino::genai ${DELAYIMP_LIB} ${CMAKE_JS_LIB})
