+++ src/python/py_vlm_pipeline.cpp
@@ -14,9 +14,11 @@
 #include "openvino/genai/visual_language/perf_metrics.hpp"
 #include "tokenizer/tokenizers_path.hpp"
 #include "py_utils.hpp"
+#include "bindings_utils.hpp"
 
 namespace py = pybind11;
 namespace pyutils = ov::genai::pybind::utils;
+namespace common_utils = ov::genai::common_bindings::utils;
 
 
 auto vlm_generate_docstring = R"(
@@ -139,7 +141,7 @@ void init_vlm_pipeline(py::module_& m) {
     py::class_<ov::genai::VLMRawPerfMetrics>(m, "VLMRawPerfMetrics", raw_perf_metrics_docstring)
         .def(py::init<>())
         .def_property_readonly("prepare_embeddings_durations", [](const ov::genai::VLMRawPerfMetrics& rw) {
-            return pyutils::get_ms(rw, &ov::genai::VLMRawPerfMetrics::prepare_embeddings_durations);
+            return common_utils::get_ms(rw, &ov::genai::VLMRawPerfMetrics::prepare_embeddings_durations);
         });
 
     py::class_<ov::genai::VLMPerfMetrics, ov::genai::PerfMetrics>(m, "VLMPerfMetrics", perf_metrics_docstring)
